+++ app/components/TracesItem.js
@@ -68,7 +68,7 @@ class TracesItem extends React.PureComponent {
       case 'uploading':
         return 'Uploading...'
       case 'uploaded':
-        return 'Uploaded'
+        return `Uploaded at ${formatDate(item.uploadedAt)}`
       case 'pending':
         if (item.errors.length === 0) {
           return 'Waiting for network...'
@@ -79,6 +79,8 @@ class TracesItem extends React.PureComponent {
           return error.message
         }
         break
+      case 'pending edit':
+        return 'Waiting to upload changes...'
       default:
         return 'An unknown error has occured'
     }