+++ app/reducers/map.js
@@ -26,7 +26,7 @@ export const initialState = {
   serialNumber: 0,
   style: style,
   selectedPhotos: false,
-  nodes: {}
+  nodes: {} // collection of member nodes
 }
 
 const TILE_CACHE_SIZE = Config.TILE_CACHE_SIZE || 10000